+++ apps/aesophia_http/src/aesophia_http_handler.erl
@@ -211,6 +211,43 @@ handle_request('ValidateByteCode', Req, _Context) ->
         _ -> {400, [], bad_request()}
     end;
 
+handle_request('GetCompilerVersion', Req, _Context) ->
+    case Req of
+        #{'ByteCodeInput' := #{<<"bytecode">> := EncodedByteCode}} ->
+            case aeser_api_encoder:safe_decode(contract_bytearray, EncodedByteCode) of
+                {ok, ByteCode} ->
+                    try
+                        Map = aeser_contract_code:deserialize(ByteCode),
+                        CVer = maps:get(compiler_version, Map, undefined),
+                        {200, [], #{version => iolist_to_binary(io_lib:format("~s", [CVer]))}}
+                    catch _:_ ->
+                        {400, [], mk_error_msg(<<"Bad bytecode">>)}
+                    end;
+                {error, _} ->
+                    {400, [], mk_error_msg(<<"Bad bytecode">>)}
+            end;
+        _ -> {400, [], bad_request()}
+    end;
+
+handle_request('GetFateAssemblerCode', Req, _Context) ->
+    case Req of
+        #{'ByteCodeInput' := #{<<"bytecode">> := EncodedByteCode}} ->
+            case aeser_api_encoder:safe_decode(contract_bytearray, EncodedByteCode) of
+                {ok, ByteCode} ->
+                    try
+                        Map = aeser_contract_code:deserialize(ByteCode),
+                        FateCode = aeb_fate_code:deserialize(maps:get(byte_code, Map)),
+                        Asm  = io_lib:format("~s", [aeb_fate_asm:pp(FateCode)]),
+                        {200, [], #{'fate-assembler' => iolist_to_binary(Asm)}}
+                    catch _:_ ->
+                        {400, [], mk_error_msg(<<"Bad bytecode">>)}
+                    end;
+                {error, _} ->
+                    {400, [], mk_error_msg(<<"Bad bytecode">>)}
+            end;
+        _ -> {400, [], bad_request()}
+    end;
+
 handle_request('Version', _Req, _Context) ->
     case aeso_compiler:version() of
         {ok, Vsn} ->