+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Added the endpoint `fate-assembler`
+- Added the endpoint `compiler-version`
 ### Changed
 ### Removed